requirements.txt
-i https://pypi.org/simple
alembic==1.13.2; python_version >= '3.8'
aniso8601==9.0.1
bcrypt==4.2.0; python_version >= '3.7'
blinker==1.8.2; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
flask==3.0.3; python_version >= '3.8'
flask-bcrypt==1.0.1
flask-cors==4.0.1
flask-login==0.6.3; python_version >= '3.7'
flask-mail==0.10.0; python_version >= '3.8'
flask-migrate==4.0.7; python_version >= '3.6'
flask-restful==0.3.10
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.0.3; python_version < '3.13' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==23.0.0; python_version >= '3.7'
honcho==1.1.0
importlib-metadata==8.2.0; python_version < '3.10'
importlib-resources==6.4.2; python_version < '3.9'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
mako==1.3.5; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
packaging==24.1; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.32; python_version >= '3.7'
sqlalchemy-serializer==1.4.12
typing-extensions==4.12.2; python_version >= '3.8'
werkzeug==3.0.3; python_version >= '3.8'
zipp==3.20.0; python_version >= '3.8'
